_ARCHIVE/image_processing/requirements.txt
opencv-python
pillow
pillow-heif
numpy
mediapipe 